backend/requirements.txt
# --- Web Framework ---
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.6
python-dotenv==1.0.1

# --- LangChain + LLM ---
langchain==0.2.0
langchain-core==0.2.0
langchain-community==0.2.0
langchain-google-genai==1.0.1
google-generativeai==0.5.2

# --- Embeddings & Vector DB ---
chromadb==0.5.0
sentence-transformers==2.7.0
huggingface-hub==0.23.2

# --- Authentication (optional) ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# --- TTS & Audio ---
cartesia==0.1.3
pydub==0.25.1

# --- ML / Transformer Models ---
torch==2.2.0
transformers==4.35.0

# --- Pydantic v2 Config ---
pydantic>=2.5.0
pydantic-settings==2.2.1
